Time: 3 ms (67.24%), Space: 5.9 MB (30.45%) - LeetHub

+++ 29-divide-two-integers/29-divide-two-integers.cpp
@@ -9,7 +9,7 @@ public:
         unsigned int b = abs(divisor);
         unsigned int ans = 0;
         while(a >= b){
-            short q = 0;
+            int q = 0;
             
             while(a > (b<<q+1)){
                 q++;
